+++ src/main/java/com/imsweb/algorithms/Algorithms.java
@@ -25,11 +25,12 @@ import com.imsweb.algorithms.iccc.IcccAlgorithm;
 import com.imsweb.algorithms.lymphoma.LymphoidNeoplasmRecodeAlgorithm2021;
 import com.imsweb.algorithms.napiia.NapiiaAlgorithm;
 import com.imsweb.algorithms.nhia.NhiaAlgorithm;
-import com.imsweb.algorithms.prcdauiho.PrcdaUihoAlgorithm;
+import com.imsweb.algorithms.prcda.PrcdaAlgorithm;
 import com.imsweb.algorithms.ruralurban.RuralUrbanAlgorithm;
 import com.imsweb.algorithms.seersiterecode.SeerSiteRecodeAlgorithm;
 import com.imsweb.algorithms.survival.SurvivalTimeAlgorithm;
 import com.imsweb.algorithms.tractestcongressdist.TractEstCongressDistAlgorithm;
+import com.imsweb.algorithms.uiho.UihoAlgorithm;
 import com.imsweb.algorithms.yostacspoverty.YostAcsPovertyAlgorithm;
 
 import static com.imsweb.algorithms.AlgorithmField.DATA_LEVEL_PATIENT;
@@ -60,7 +61,8 @@ public final class Algorithms {
     public static final String ALG_ICCC = "iccc";
     public static final String ALG_IARC = "iarc-multiple-primary";
     public static final String ALG_COUNTY_AT_DIAGNOSIS_ANALYSIS = "county-at-diagnosis-analysis";
-    public static final String ALG_PRCDA_UIHO = "prcda-uiho";
+    public static final String ALG_PRCDA = "prcda";
+    public static final String ALG_UIHO = "uiho";
     public static final String ALG_ACS_LINKAGE = "acs-linkage";
     public static final String ALG_EPHT_SUBCOUNTY = "epht-subcounty";
     public static final String ALG_TRACT_EST_CONGRESS_DIST = "tract-est-congress-dist";
@@ -153,7 +155,9 @@ public final class Algorithms {
     public static final String FIELD_IARC_MP_HISTOLOGY = "iarcMpHistologicTypeIcdO3";
     public static final String FIELD_COUNTY_AT_DX_ANALYSIS_FLAG = "countyAtDxAnalysisFlag";
     public static final String FIELD_IHS_PRCDA = "ihsPurchRefCareDeliveryArea";
+    public static final String FIELD_IHS_PRCDA_2017 = "ihsPurchRefCareDeliveryArea2017";
     public static final String FIELD_UIHO = "urbanIndianHealthOrganization";
+    public static final String FIELD_UIHO_CITY = "uihoFacility";
     public static final String FIELD_ACS_YOST_QUINTILE_US = "yostQuintileUS";
     public static final String FIELD_ACS_YOST_QUINTILE_STATE = "yostQuintileState";
     public static final String FIELD_ACS_POV_ALL_RACES = "acsPctPovAllRaces";
@@ -233,6 +237,9 @@ public final class Algorithms {
             addField(AlgorithmField.of(FIELD_CENSUS_2000, 130, 6, "Census Tract 2000", "Cens 2000", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_CENSUS_2010, 135, 6, "Census Tract 2010", "Cens 2010", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_CENSUS_POVERTY_INDICTR, 145, 1, "Census Tr Poverty Indictr", "Cens Pov", DATA_LEVEL_TUMOR));
+            addField(AlgorithmField.of(FIELD_IHS_PRCDA, 194, 1, "IHS Purchased/Referred Care Delivery Area", "IHS PRCDA", DATA_LEVEL_TUMOR));
+            addField(AlgorithmField.of(FIELD_UIHO, 284, 1, "Urban Indian Health Organization (UIHO)", "UIHO", DATA_LEVEL_TUMOR));
+            addField(AlgorithmField.of(FIELD_UIHO_CITY, 285, 2, "UIHO City", "UIHO City", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_URIC_2000, 345, 1, "URIC 2000", "URIC 2000", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_URIC_2010, 346, 1, "URIC 2010", "URIC 2010", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_RUCA_2000, 339, 1, "RUCA 2000", "RUCA 2000", DATA_LEVEL_TUMOR));
@@ -279,8 +286,7 @@ public final class Algorithms {
             addField(AlgorithmField.of(FIELD_IARC_MP_HIST_GROUP, null, 2, "IARC Multiple Primary Histology Group", "IARC MP Hist Grp", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_IARC_MP_HISTOLOGY, null, 4, "IARC Multiple Primary Histology", "IARC MP Hist", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_COUNTY_AT_DX_ANALYSIS_FLAG, null, 4, "County at DX Analysis Flag", "DX Anlys Cty Fg", DATA_LEVEL_TUMOR));
-            addField(AlgorithmField.of(FIELD_IHS_PRCDA, null, 1, "IHS Purchased/Referred Care Delivery Area (PRCDA)", "IHS PRCDA", DATA_LEVEL_TUMOR));
-            addField(AlgorithmField.of(FIELD_UIHO, null, 1, "Urban Indian Health Organization (UIHO)", "UIHO", DATA_LEVEL_TUMOR));
+            addField(AlgorithmField.of(FIELD_IHS_PRCDA_2017, null, 1, "IHS PRCDA 2017", "IHS PRCDA 2017", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_ACS_YOST_QUINTILE_US, null, 1, "Yost Quintile (Across States)", "Yost US", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_ACS_YOST_QUINTILE_STATE, null, 1, "Yost Quintile (State Specific)", "Yost State", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_ACS_POV_ALL_RACES, null, 6, "ACS Pct Poverty (All Races)", "ACS Pov All", DATA_LEVEL_TUMOR));
@@ -312,7 +318,8 @@ public final class Algorithms {
             addAlgorithm(new IcccAlgorithm());
             addAlgorithm(new IarcAlgorithm());
             addAlgorithm(new CountyAtDxAnalysisAlgorithm());
-            addAlgorithm(new PrcdaUihoAlgorithm());
+            addAlgorithm(new PrcdaAlgorithm());
+            addAlgorithm(new UihoAlgorithm());
             addAlgorithm(new YostAcsPovertyAlgorithm());
             addAlgorithm(new EphtSubCountyAlgorithm());
             addAlgorithm(new TractEstCongressDistAlgorithm());