+++ src/main/java/com/imsweb/algorithms/prcda/PrcdaOutputDto.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2013 Information Management Services, Inc.
+ */
+package com.imsweb.algorithms.prcda;
+
+public class PrcdaOutputDto {
+
+    private String _prcda;
+    private String _prcda2017;
+
+    public String getPrcda() {
+        return _prcda;
+    }
+
+    public String getPrcda2017() {
+        return _prcda2017;
+    }
+
+    public void setPrcda(String str) {
+        _prcda = str;
+    }
+
+    public void setPrcda2017(String str) {
+        _prcda2017 = str;
+    }
+}